chore: refactor constants into a shared file
Moves `src/pages/Background/constants.ts` to `src/constants.ts` to consolidate and avoid duplicating string constants like `LOADING_COPY` and `SIGNED_OUT_CTA_COPY` in `src/pages/Content/script.ts` and others.
